should fail on test failure

+++ .circleci/test.sh
@@ -7,6 +7,10 @@ echo "" > testreport.txt
 for d in $(go list ./... | grep -v vendor); do
   echo Testing: $d
   CGO_ENABLED=0 go test -short -v -coverprofile=profile.out -covermode=atomic $d 2>&1 > test.out
+  if [ "$?" -ne "0" ]; then
+    cat test.out
+    exit $?
+  fi
   if [ -f profile.out ]; then
     cat profile.out >> coverage.txt
     rm profile.out
